Tmux: fix issue where status line is messed up
Issue
------
When Tmux displays/refreshes the status line it scrolls upwards whereas
it should not.
The issue occurs when tmux-online-status plugin displays the icons
for macOS in iTerm2 (Build 3.1.6beta3).
I have not experienced this issue in iTerm2 *stable*, nor on Linux.

Solution
---------
Sticking to the icons for Linux fixes things up.

Context
-------
- MacOS: version 10.13.3
- iTerm2: Build 3.1.6beta3