Make sure openssl installed by ruby-build is updated
ruby-build installs a non-Homebrew OpenSSL for each Ruby version
installed and these are never upgraded.

To link Rubies ruby-build installs a non-Homebrew OpenSSL for each Ruby
version installed and these are never upgraded.

To link Rubies to Homebrew's OpenSSL 1.1 (which is upgraded)
add the following to your ~/.zshrc:
  export RUBY_CONFIGURE_OPTS="--with-openssl-dir=$(brew --prefix openssl@1.1)"

Note: this may interfere with building old versions of Ruby (e.g <2.4) that use
OpenSSL <1.1.
Already up-to-date.